+++ multiboot2/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - dependency updates
 - **Breaking:** MSRV is now 1.75
+- Added missing tags:
+    - `ApmTag`
+    - `BootdevTag`
+    - `NetworkTag`
 
 ## v0.22.2 (2024-08-24)