+++ multiboot2/CHANGELOG.md
@@ -4,10 +4,11 @@
 
 - dependency updates
 - **Breaking:** MSRV is now 1.75
-- Added missing tags:
+- Added missing tags along with getters for on `BootInformation`:
     - `ApmTag`
     - `BootdevTag`
     - `NetworkTag`
+- `BootInformation::tags` iterator is now public
 
 ## v0.22.2 (2024-08-24)