+++ multiboot2/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v0.23.0 (2024-09-17)
+
 - dependency updates
 - **Breaking:** MSRV is now 1.75
 - Added missing tags along with getters for on `BootInformation`: